Put {enable,disable}PeriodicUpdatesReceiver back
This is needed after all. It seemed to work even without this, but we
were receiving the updates all the time even when the addon was supposed
to be closed, and it's likely this was a one time thing on my phone and
it wouldn't work at all for others.